Update README.md
A to-do list is a list of tasks you need to complete or things you want to do1,2 3. The main purpose of a to-do list is to provide yourself with a list of your priorities to ensure that you don't forget anything and can effectively plan out your tasks so that they are all accomplished in the correct time frame.